ml_api/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
joblib==1.3.2

# Database & External APIs
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0

# Utilities
pydantic==2.5.0
httpx==0.25.2

# Date/Time utilities
python-dateutil==2.8.2
jpholiday==0.1.9  # Japanese holidays